requirements.txt
boa-api>=0.1.13
jinja2>=3.1.2
jsonschema>=3.2.0
pandas>=1.4.2
pyarrow>=5.0.0
python-dotenv>=0.21.0
requests>=2.28.1
